requirements.txt
python-telegram-bot==9.0.0
requests==2.18.4
wolframalpha==3.0
wit==4.3.0
ffmpy==0.2.2
yandex.translate==0.3.5